refactor: replace permission handler with PermissionRequestProcessor
